Re #4224: Relevant parts of simpgmspace.cpp moved into simufatfs.cpp and simudisk.cpp
